feat(onboarding using R&D with bertrand boilerplate)
